+++ src/assetbundles/bugsnag/BugsnagAsset.php
@@ -11,7 +11,7 @@ class BugsnagAsset extends AssetBundle
     // Public Methods
     // =========================================================================
 
-    public function init()
+    public function init(): void
     {
         $this->depends = [
             VerbbCpAsset::class,